Docs update (#71)
Main changes:

* Add page documenting improvements over the original model, to comply with JMLR-MLOSS' requirements.

* Added a `load_random_evotuning_params` function, which allows us to start from a random starting point.

* Updated docs with copy/paste-able example of how to do evotuning.

-----

The rest of the changes below

* Adding latex build files

* Fixed markdown string violation

* Add comment to params

* Code formatting

* Run style checks before running tests.

This will reduce frustrations waiting for tests to finish
only to find that it was a trivial code formatting
that cause tests to fail.

* Changed Makefile commands

* Reinstated black -l 79 .

I took it out because I thought the `pyproject.toml` would be respected in the CI.
Turns out not.

* Fixed code with updated versions of black and isort

* Made provider "pages" instead of "pages:git".

This should be in-line with the latest API.

* Add test for loading random evotuning params

* Added back some imports to make copy/pasting even easier.

* Updated improvements page with comment on code coverage

* Added link to preprint