Embed sequences while evotuning. (#76)
This solves the problem with large rounds of evotuning, where we have to worry about memory storage of large tensors.
In the short-run, we expect evotuning on smaller, artisanal batches of sequences (sorry!) to be slower, but there may be a neat workaround later on that can be included, such as asynchronously generating the next batch using a worker process always waiting in the background while the gradient update steps happen.

Apart from that, there were other minor changes, including more explicit debugging outputs from Makefile commands, which should make debugging easiser in the future.

Co-authored-by: Eric Ma <ericmajinglong@gmail.com>