Shifted step layer inside scope of batch function. (#74)
This avoids creating a new function object on each call of mLSTM1900_batch, which might be the reason why we have
memory leakage issues, as the new object doesn't hit lax.scan's function cache.

We have thus sacrificed unit test-ability of the step function, but at least we can still test it from the outside. Not that we'd want to modify the step function anyways.